+++ packages/web/src/modules/calendar/components/TimerState.vue
@@ -100,7 +100,7 @@ syncTimerState(isStarted.value);
     </ly-button>
   </div>
   <ly-modal v-model="showTimerEdit" title="timer.edit.title" @submit="submitEditTimer">
-    <ly-input-time-number v-model="timerEditValue" :max="max" />
+    <ly-input-time-number v-model="timerEditValue" :max="max || undefined" />
   </ly-modal>
 </template>