+++ CHANGELOG.md
@@ -6,13 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [0.1.0.alpha2] - WIP
+## [0.1.0.alpha2] - 2023-05-11
 ### Fixed
 - User is not redirected to newly created profile
 - Email is not marked as verified after registration
 - Handling of avatars in multi-user profiles
 - Slider navigation overflow visible
 - Timer datapoint value can not be updated manually
+- Datapoint timer input overflows on small devices
+- Manual timer input not working if max value is set to 0
 
 ### Changed
 - Added missing coloring of some error alerts